[Add] Loot chests! (pls read)
There's the closed and opened chests, and yes, they are two different objects with two different UVs and textures, but I got it close enough. They are both set to the origin, so just swap em when the thing is opened; eventually I'll fix it so they're both using the same UVs and textures (except for the lights), but expo. Either way, pop the lootChest-contents.obj at the same location to have the chest have stuff in it. For the texture of lootChest-contents.obj just use Sednium's textures, considering it's the same thing. And I've added nicer Sednium textures.