[Fix] Did a lot in the textures and models folder [READ DESCRIPTION]
Y'all ready for this

This shit is gonna require a lot of fixing in the code, and I'm very sorry, but this makes life easier, trust. 

Basically what I did was adding folders to sort all the textures that I have, renaming them in the process to keep the names consistent with PBR naming standards. Along with this, I renamed the semi-auto rifle to the marksman rifle, just to make it sound a bit better. Oh, and it includes the new fixed textures for the chests, the last push had the lights on the back.