[Add] HUD Overhaul!
Added in the new HUD (primordial-hud-overlay.png) as well as two examples of its use just in case, with one being with the primary equipped and one with the secondary weapon equipped. Also, you'll see five .pngs that are the weapon silhouettes. These are to be used for each respective character, but also for the character select to indicate which weapons they have since that was a requested feature.