[Add] Guns bank
All sounds that I had (should be all of them that are in at this point) have now been put into banks, so if you merge from this and aren't using the FMOD stuff, don't expect to hear much of anything. 

All of the gun sounds have subtle pitch modulation to vary up the soundscape a little bit, so that's neat. Will probably do this with the footsteps whenever I can be fucked to put that in.