[Add] COMPLETE SONG - READ DESCRIPTION PLS
In the Music bank (obviously), and the event/track is titled "Ancient". That's the gameplay one, and if you want a shitty gameplay one, that's "Unyielding". For the main menu, it's called "Derelict", and all you need to do is set the parameter levelClear to 1 in the character select or something to end the song. Both gameplay tracks have levelClear defaulted to 0 as well, and setting that to 1 will end the track, with the other parameter inCombat (also defaulted to 0) set to 1 for the combat loop. This value doesn't effect whether or not the song ends when levelClear is 0.